requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
discord.py==0.16.12
idna==2.6
mongoengine==0.15.0
multidict==3.3.2
munch==2.2.0
pymongo==3.6.0
pyyaml==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
websockets==3.4
